chore: forzar lint + type-check + tests + CI en cada commit/push
- mypy --strict ahora corre en pre-commit (via uv run, vee el venv
  del proyecto con todas las deps tipadas). Cero errores de tipo.
- pytest full suite corre como hook pre-push. Bloquea git push si
  fallan tests. No traba commits individuales (suite es lenta).
- Workflow .github/workflows/ci.yml: ruff + ruff-format + mypy +
  pytest contra Postgres 16. Es la unica forma de impedir que un
  commit con --no-verify llegue a main.
- Overrides de mypy para tests/*: disallow_untyped_defs=false
  (fixtures con tipos complejos no anotables sin ruido).
- Fix tipado del bridge_client: __enter__ retorna Self (BridgeClient),
  get_checkpoint/put_checkpoint mantienen signatura del parent
  (source_id: int) y descartan el arg con del — sin Liskov violations.

+++ src/memex_local/scheduler.py
@@ -85,7 +85,7 @@ def _build_runtimes(
 
 def _backoff_seconds(failures: int) -> float:
     """Backoff exponencial con techo de 1h."""
-    return min(60 * (2 ** min(failures, 6)), 3600)
+    return float(min(60 * (2 ** min(failures, 6)), 3600))
 
 
 class Scheduler:
